REFACTOR: Sinav ve Ders Porgrami icin yazilan module ve class adlari degistirildi.
