+++ ulakbus/models/__init__.py
@@ -13,5 +13,4 @@ from .ogrenci import *
 from .hitap.hitap import *
 from .buildings_rooms import *
 from .form import *
-from .ders_programi import *
-from .ders_programi_data import *
+from .ders_sinav_programi import *
